+++ instrumentation/opentelemetry-instrumentation-botocore/src/opentelemetry/instrumentation/botocore/extensions/__init__.py
@@ -33,6 +33,7 @@ def _lazy_load(module, cls):
 
 _KNOWN_EXTENSIONS = {
     "dynamodb": _lazy_load(".dynamodb", "_DynamoDbExtension"),
+    "lambda": _lazy_load(".lmbd", "_LambdaExtension"),
     "sqs": _lazy_load(".sqs", "_SqsExtension"),
 }